+++ Runtime/AssemblyInfo.cs
@@ -4,6 +4,7 @@ using System.Runtime.CompilerServices;
 [assembly: AssemblyVersion("2.0.2.0")]
 [assembly: AssemblyFileVersion("2.0.2.0")]
 [assembly: AssemblyInformationalVersion("2.0.2-development")]
+[assembly: InternalsVisibleTo("YarnSpinner.Editor.Tests")]
 [assembly: InternalsVisibleTo("YarnSpinnerTests")]
 [assembly: InternalsVisibleTo("YarnSpinner.Unity.Editor")]